application_server/requirements.txt
Django==3.0.5
imutils==0.4.5
numpy==1.14.5
opencv-python==4.2.0.34
Pillow==7.1.2
pytz==2017.3
six==1.10.0
tornado==4.2
PyMySQL==0.9.3
